code add for authorization

+++ backend/routes/posts.js
@@ -37,8 +37,10 @@ router.post(
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath : url + "/images/" + req.file.filename
+        imagePath : url + "/images/" + req.file.filename,
+        creator: req.userData.userId
     });
+
     post.save().then(createdPost => {
         res.status(201).json({
             message: "Post added successfully",
@@ -64,13 +66,20 @@ router.put(
         _id: req.params.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
     
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        res.status(200).json({
-            message: "Post update successfully",
-        })
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+        if(result.modifiedCount > 0) {
+            res.status(200).json({
+                message: "Post update successfully",
+            })
+        } else {
+            res.status(400).json({
+                message: "Not authorized!",
+            })
+        }
     });
 })
 
@@ -111,10 +120,16 @@ router.delete(
     "/:id", 
     checkAuth,
     (req,res, next)=> {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        res.status(200).json({
-            message: "Deleted successfully"
-        })
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        if(result.deletedCount > 0) {
+            res.status(200).json({
+                message: "Post delete successfully",
+            })
+        } else {
+            res.status(400).json({
+                message: "Not authorized!",
+            })
+        }
     })
 })
